requirements.txt
pandas==0.25.0
numpy==1.17.0
tensorflow==1.14.0
tensorflow-hub==0.5.0
PyYAML==5.1.2
pyyml==0.0.2
gcsfs==0.3.1
